Move encode/decode to dedicated module
